refactor(price): extract currency model + fiat flag to aggregates/currency-select
Move user-preference and orchestration logic (currencyModel, priceProviderModel)
out of domains/price into aggregates/currency-select, keeping the domain purely
for data-fetching and types. This follows the same aggregate pattern as
wallet-select and staking-network.
