+++ src/renderer/aggregates/currency-select/index.ts
@@ -1,0 +1,1 @@
+export { currencySelect } from './model';
