+++ src/renderer/features/assets/AssetsChainView/ui/NetworkFiatBalance.tsx
@@ -5,7 +5,8 @@ import { useEffect, useState } from 'react';
 import { type Asset, type Balance } from '@/shared/core';
 import { useI18n } from '@/shared/i18n';
 import { formatFiatBalance, getRoundedValue, totalAmount } from '@/shared/lib/utils';
-import { currencyModel, priceProviderModel, useAssetsPrices } from '@/domains/price';
+import { useAssetsPrices } from '@/domains/price';
+import { currencySelect } from '@/aggregates/currency-select';
 import { FiatBalance } from '@/widgets/price';
 
 type Props = {
@@ -18,10 +19,10 @@ export const NetworkFiatBalance = ({ assets, balances, className }: Props) => {
   const { t } = useI18n();
   const [fiatAmount, setFiatAmount] = useState<BigNumber>(new BigNumber(0));
 
-  const currency = useUnit(currencyModel.$activeCurrency);
-  const pricesParams = useUnit(priceProviderModel.$currentPricesParams);
+  const currency = useUnit(currencySelect.$activeCurrency);
+  const pricesParams = useUnit(currencySelect.$currentPricesParams);
   const { data: prices } = useAssetsPrices(pricesParams);
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
+  const fiatFlag = useUnit(currencySelect.$fiatFlag);
 
   useEffect(() => {
     // TODO: Move logic to model https://app.clickup.com/t/8692tr8x0
