+++ src/renderer/features/assets/AssetsPortfolioView/ui/AssembledAssetAmount.tsx
@@ -7,7 +7,7 @@ import { useI18n } from '@/shared/i18n';
 import { cnTw } from '@/shared/lib/utils';
 import { AssetBalance } from '@/shared/ui-entities';
 import { Skeleton, Tooltip } from '@/shared/ui-kit';
-import { priceProviderModel } from '@/domains/price';
+import { currencySelect } from '@/aggregates/currency-select';
 import { AssetFiatBalance } from '@/widgets/price';
 
 type Props = PropsWithChildren & {
@@ -18,7 +18,7 @@ type Props = PropsWithChildren & {
 export const AssembledAssetAmount = ({ balance, asset }: Props) => {
   const { t } = useI18n();
 
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
+  const fiatFlag = useUnit(currencySelect.$fiatFlag);
 
   if (!balance) {
     return (
