+++ src/renderer/features/currency/CurrencyForm/model/currency-form.ts
@@ -3,7 +3,7 @@ import { createForm } from 'effector-forms';
 import { combineEvents, spread } from 'patronum';
 
 import { type CurrencyItem } from '@/domains/price';
-import { currencyModel, priceProviderModel } from '@/domains/price';
+import { currencySelect } from '@/aggregates/currency-select';
 
 export type Callbacks = {
   onSubmit: () => void;
@@ -24,13 +24,13 @@ const $currencyForm = createForm({
   validateOn: ['submit'],
 });
 
-const $cryptoCurrencies = currencyModel.$currencyConfig.map((config) => {
+const $cryptoCurrencies = currencySelect.$currencyConfig.map((config) => {
   return config.filter((c) => c.category === 'crypto');
 });
-const $popularFiatCurrencies = currencyModel.$currencyConfig.map((config) => {
+const $popularFiatCurrencies = currencySelect.$currencyConfig.map((config) => {
   return config.filter((c) => c.category === 'fiat' && c.popular);
 });
-const $unpopularFiatCurrencies = currencyModel.$currencyConfig.map((config) => {
+const $unpopularFiatCurrencies = currencySelect.$currencyConfig.map((config) => {
   return config.filter((c) => c.category === 'fiat' && !c.popular);
 });
 
@@ -45,8 +45,8 @@ const $isFormValid = combine(
 sample({
   clock: formInitiated,
   source: {
-    fiatFlag: priceProviderModel.$fiatFlag,
-    currency: currencyModel.$activeCurrency,
+    fiatFlag: currencySelect.$fiatFlag,
+    currency: currencySelect.$activeCurrency,
   },
   fn: ({ fiatFlag, currency }) => ({ fiatFlag, currency: currency?.id || 0 }),
   target: $currencyForm.setInitialForm,
@@ -60,15 +60,15 @@ sample({
   },
   target: spread({
     targets: {
-      fiatFlag: priceProviderModel.events.fiatFlagChanged,
-      currency: currencyModel.events.currencyChanged,
+      fiatFlag: currencySelect.events.fiatFlagChanged,
+      currency: currencySelect.events.currencyChanged,
     },
   }),
 });
 
 sample({
   clock: combineEvents({
-    events: [priceProviderModel.events.fiatFlagChanged, currencyModel.output.currencyChanged],
+    events: [currencySelect.events.fiatFlagChanged, currencySelect.output.currencyChanged],
   }),
   target: attach({
     source: $callbacks,
