+++ src/renderer/features/dashboard-portfolio-overview/hooks/useBalanceAllocation.ts
@@ -3,9 +3,10 @@ import { useUnit } from 'effector-react';
 import { useMemo } from 'react';
 
 import { getRoundedValue, transferableAmountBN } from '@/shared/lib/utils';
-import { currencyModel, priceProviderModel, useAssetsPrices } from '@/domains/price';
+import { useAssetsPrices } from '@/domains/price';
 import { balanceModel } from '@/entities/balance';
 import { networkModel } from '@/entities/network';
+import { currencySelect } from '@/aggregates/currency-select';
 
 export type AllocationData = {
   transferablePct: number;
@@ -16,8 +17,8 @@ export type AllocationData = {
 export const useBalanceAllocation = (accountIds: string[]): AllocationData | null => {
   const balanceMap = useUnit(balanceModel.$balanceMap);
   const chains = useUnit(networkModel.$chains);
-  const currency = useUnit(currencyModel.$activeCurrency);
-  const pricesParams = useUnit(priceProviderModel.$currentPricesParams);
+  const currency = useUnit(currencySelect.$activeCurrency);
+  const pricesParams = useUnit(currencySelect.$currentPricesParams);
   const { data: prices } = useAssetsPrices(pricesParams);
 
   return useMemo(() => {
