+++ src/renderer/features/dashboard-portfolio-overview/hooks/useChainBreakdown.ts
@@ -4,9 +4,10 @@ import { useMemo } from 'react';
 
 import { type ChainId } from '@/shared/core';
 import { getRoundedValue, totalAmount, totalAmountBN } from '@/shared/lib/utils';
-import { currencyModel, priceProviderModel, useAssetsPrices } from '@/domains/price';
+import { useAssetsPrices } from '@/domains/price';
 import { balanceModel } from '@/entities/balance';
 import { networkModel } from '@/entities/network';
+import { currencySelect } from '@/aggregates/currency-select';
 
 export type ChainAssetRow = {
   assetId: number;
@@ -30,8 +31,8 @@ export type ChainBreakdownData = {
 export const useChainBreakdown = (chainId: ChainId, accountIds: string[]): ChainBreakdownData => {
   const balanceMap = useUnit(balanceModel.$balanceMap);
   const chains = useUnit(networkModel.$chains);
-  const currency = useUnit(currencyModel.$activeCurrency);
-  const pricesParams = useUnit(priceProviderModel.$currentPricesParams);
+  const currency = useUnit(currencySelect.$activeCurrency);
+  const pricesParams = useUnit(currencySelect.$currentPricesParams);
   const { data: prices } = useAssetsPrices(pricesParams);
 
   return useMemo(() => {
