+++ src/renderer/features/dashboard-staking-overview/hooks/useRewardsBreakdown.ts
@@ -3,8 +3,9 @@ import { useUnit } from 'effector-react';
 import { useMemo } from 'react';
 
 import { getRoundedValue } from '@/shared/lib/utils';
-import { currencyModel, priceProviderModel, useAssetsPrices } from '@/domains/price';
+import { useAssetsPrices } from '@/domains/price';
 import { type RewardsMap } from '@/domains/staking';
+import { currencySelect } from '@/aggregates/currency-select';
 
 import { type EntryLike } from './useStakingBreakdown';
 import { type ChainRewardsSummary } from './useTotalRewards';
@@ -31,8 +32,8 @@ type Params = {
 };
 
 export const useRewardsBreakdown = ({ rewardsMap, chainSummary, accountIds, allEntries }: Params) => {
-  const currency = useUnit(currencyModel.$activeCurrency);
-  const pricesParams = useUnit(priceProviderModel.$currentPricesParams);
+  const currency = useUnit(currencySelect.$activeCurrency);
+  const pricesParams = useUnit(currencySelect.$currentPricesParams);
   const { data: prices } = useAssetsPrices(pricesParams);
 
   return useMemo(() => {
