+++ src/renderer/features/staking-bond-nominate/components/BondFormShards.tsx
@@ -10,8 +10,8 @@ import { type RadioOption } from '@/shared/ui/types';
 import { AssetBalance, Identicon } from '@/shared/ui-entities';
 import { Tooltip } from '@/shared/ui-kit';
 import { useAccountsNames } from '@/domains/network';
-import { priceProviderModel } from '@/domains/price';
 import { AccountAddress, ProxyWalletAlert, accountUtils, walletUtils } from '@/entities/wallet';
+import { currencySelect } from '@/aggregates/currency-select';
 import { walletSelect } from '@/aggregates/wallet-select';
 import { AmountInput } from '@/features/assets-balances';
 import { AssetFiatBalance } from '@/widgets/price';
@@ -271,7 +271,7 @@ const FeeSection = () => {
   const isFeeLoading = useUnit(formModelShards.$isFeeLoading);
   const isMultisig = useUnit(formModelShards.$isMultisig);
 
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
+  const fiatFlag = useUnit(currencySelect.$fiatFlag);
 
   if (!network || shards.value.length === 0) {
     return null;
