+++ src/renderer/features/transfer/ui/TokenSelector.tsx
@@ -15,8 +15,9 @@ import {
 import { BodyText, EmptyListWithIcon, FootnoteText } from '@/shared/ui';
 import { AssetBalance, AssetIcon, ChainIcon } from '@/shared/ui-entities';
 import { Modal, ScrollArea, SearchInput } from '@/shared/ui-kit';
-import { currencyModel, priceProviderModel, useAssetsPrices } from '@/domains/price';
+import { useAssetsPrices } from '@/domains/price';
 import { balanceModel, balanceUtils } from '@/entities/balance';
+import { currencySelect } from '@/aggregates/currency-select';
 import { AssetFiatBalance, FiatBalance, TokenPrice } from '@/widgets/price';
 import { formModel } from '../model/form-model';
 
@@ -33,9 +34,9 @@ export const TokenSelectorModal = memo(({ isOpen, chain, onSelect, onClose }: Pr
 
   const initiator = useUnit(formModel.form.fields.initiator.$value);
   const balances = useUnit(balanceModel.$balanceMap);
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
-  const currency = useUnit(currencyModel.$activeCurrency);
-  const pricesParams = useUnit(priceProviderModel.$currentPricesParams);
+  const fiatFlag = useUnit(currencySelect.$fiatFlag);
+  const currency = useUnit(currencySelect.$activeCurrency);
+  const pricesParams = useUnit(currencySelect.$currentPricesParams);
   const { data: prices } = useAssetsPrices(pricesParams);
 
   const assetsWithBalances = useMemo(() => {
