+++ src/renderer/widgets/EditDelegationModal/shards/ui/DelegateForm.tsx
@@ -9,11 +9,11 @@ import { Button, DetailRow, FootnoteText, Icon, InputHint, SmallTitleText } from
 import { AssetBalance, SignatorySelect } from '@/shared/ui-entities';
 import { Checkbox, Modal, Tooltip } from '@/shared/ui-kit';
 import { accounts } from '@/domains/network';
-import { priceProviderModel } from '@/domains/price';
 import { balanceModel, balanceUtils } from '@/entities/balance';
 import { OperationTitle } from '@/entities/chain';
 import { BalanceDiff, LockPeriodDiff, LockValueDiff } from '@/entities/governance';
 import { ProxyWalletAlert, walletModel } from '@/entities/wallet';
+import { currencySelect } from '@/aggregates/currency-select';
 import { AmountInput } from '@/features/assets-balances';
 import { lockPeriodsModel, locksPeriodsAggregate } from '@/features/governance';
 import { ConvictionSelect } from '@/widgets/VoteModal';
@@ -232,7 +232,7 @@ const FeeSection = () => {
     fn: (locks, [chain]) => (chain ? (locks[chain.chainId] ?? null) : null),
   });
 
-  const fiatFlag = useUnit(priceProviderModel.$fiatFlag);
+  const fiatFlag = useUnit(currencySelect.$fiatFlag);
 
   useGate(locksPeriodsAggregate.gates.flow, { chain: network?.chain });
